UPDATE: REPLY JS FILE ABOUT REPLY API EXCEPTION update reply js file about endpoint reply method post, get, put, delete exception

+++ router/reply.js
@@ -21,33 +21,52 @@ router.post("/",(req,res)=>{
     const api_call_time=moment()
 
     const result={
-        "success":false
+        "success":false,
+        "message":null
     }
 
-    if(tokenVerify(token_public)){
-
-        const db = new Client(pgInit)
-        db.connect((err)=>{
-            if(err) {
-                console.log(err)
-            }
-        })
-        const sql="INSERT INTO badonnaproject.reply(coment_num,contents) VALUES($1,$2)"
-        const values=[coment_number,reply_contents]
-        
-        db.query(sql,values,(err,row)=>{
-            if(!err){
-                result.success=true
-            }else{
-                console.log(err)
-            }
-
-            //로깅 남기기
-            logFuntion(api_name,req_host, req_data, row.rows[0],api_call_time)
-
+    try{
+        if(coment_number.length == 0 || coment_number == null || Number.isInteger(coment_number)){
+            result.message="옳바르지 않은 댓글 번호 입니다."
             res.send(result)
-            db.end()
-        })
+        }else if(reply_contents.length == 0 || reply_contents == null || reply_contents >200){
+            result.message="옳바르지 않은 대댓글 내용 입니다."
+            res.send(result)
+        }else{
+
+            if(tokenVerify(token_public)){
+
+                const db = new Client(pgInit)
+                db.connect((err)=>{
+                    if(err) {
+                        console.log(err)
+                    }
+                })
+                const sql="INSERT INTO badonnaproject.reply(coment_num,contents) VALUES($1,$2)"
+                const values=[coment_number,reply_contents]
+                
+                db.query(sql,values,(err,row)=>{
+                    if(!err){
+                        result.success=true
+                    }else{
+                        console.log(err)
+                    }
+
+                    //로깅 남기기
+                    logFuntion(api_name,req_host, req_data, row.rows[0],api_call_time)
+
+                    res.send(result)
+                    db.end()
+                })
+            }else{
+                result.message="잘못된 토큰 입니다."
+                res.send(result)
+            }
+        }
+
+    }catch(e){
+        result.message="에러 입니다."
+        res.send(result)
     }
 
 
@@ -67,35 +86,50 @@ router.get("/",(req,res)=>{
 
     const result={
         "success":false,
-        "data":null
+        "data":null,
+        "message":null
     }
-    if(tokenVerify(token_public)){
 
-        const db = new Client(pgInit)
-        db.connect((err)=>{
-            if(err) {
-                console.log(err)
-            }
-        })
-        const sql="SELECT contents FROM  badonnaproject.reply WHERE reply_num=$1"
-        const values=[reply_number]
-        
-        db.query(sql,values,(err,row)=>{
-            if(!err){
-                result.success=true
-                result.data=row.rows[0]
+    try{
+        if(reply_number.length ==0 || reply_number == null || Number.isInteger(reply_number)){
+            result.message="옳바르지 않은 대댓글 번호 입니다."
+            result.send(result)
+        }else{
+
+            if(tokenVerify(token_public)){
+
+                const db = new Client(pgInit)
+                db.connect((err)=>{
+                    if(err) {
+                        console.log(err)
+                    }
+                })
+                const sql="SELECT contents FROM  badonnaproject.reply WHERE reply_num=$1"
+                const values=[reply_number]
+                
+                db.query(sql,values,(err,row)=>{
+                    if(!err){
+                        result.success=true
+                        result.data=row.rows[0]
+                    }else{
+                        console.log(err)
+                    }
+
+                    //로깅 남기기
+                    logFuntion(api_name,req_host, req_data, row.rows[0],api_call_time)
+
+                    res.send(result)
+                    db.end()
+                })
             }else{
-                console.log(err)
+                result.message="잘못된 토큰"
+                res.send(result)
             }
+        }
 
-            //로깅 남기기
-            logFuntion(api_name,req_host, req_data, row.rows[0],api_call_time)
+    }catch(e){
 
-            res.send(result)
-            db.end()
-        })
     }
-
 
 })
 
@@ -113,33 +147,55 @@ router.put("/",(req,res)=>{
 
     const result={
         "success":false,
+        "message":null
     }
-    if(tokenVerify(token_public)){
 
-        const db = new Client(pgInit)
-        db.connect((err)=>{
-            if(err) {
-                console.log(err)
-            }
-        })
-        const sql="UPDATE badonnaproject.reply SET contents=$3 WHERE coment_num=$1 AND reply_num=$2"
-        const values=[coment_number,reply_number,reply_contents]
-        
-        db.query(sql,values,(err,row)=>{
-            if(!err){
-                result.success=true
-            }else{
-                console.log(err)
-            }
+    try{
 
-            //로깅 남기기
-            logFuntion(api_name,req_host, req_data, row.rows[0],api_call_time)
-
+        if(coment_number.length == 0 || coment_number == null || Number.isInteger(coment_number)){
+            result.message="옳바르지 않은 댓글 번호 입니다."
             res.send(result)
-            db.end()
-        })
-    }
+        }else if(reply_number.length == 0 || reply_number == null || Number.isInteger(reply_number)){
+            result.message="옳바르지 않은 대댓글 번호 입니다."
+            res.send(result)
+        }else if(reply_contents.length == 0 || reply_contents == null || reply_contents.length >200){
+            result.message="옳바르지 않은 대댓글 내용 입니다."
+            res.send(result)
+        }else{
+            if(tokenVerify(token_public)){
 
+                const db = new Client(pgInit)
+                db.connect((err)=>{
+                    if(err) {
+                        console.log(err)
+                    }
+                })
+                const sql="UPDATE badonnaproject.reply SET contents=$3 WHERE coment_num=$1 AND reply_num=$2"
+                const values=[coment_number,reply_number,reply_contents]
+                
+                db.query(sql,values,(err,row)=>{
+                    if(!err){
+                        result.success=true
+                    }else{
+                        console.log(err)
+                    }
+
+                    //로깅 남기기
+                    logFuntion(api_name,req_host, req_data, row.rows[0],api_call_time)
+
+                    res.send(result)
+                    db.end()
+                })
+            }else{
+                result.message="잚못된 토큰"
+                res.send(result)
+            }
+        }
+
+    }catch(e){
+        result.message="에러 입니다."
+        res.send(result)
+    }
 })
 
 
@@ -154,36 +210,52 @@ router.delete("/",(req,res)=>{
     const res_data=" "
 
     const result={
-        "success":false
+        "success":false,
+        "message":null
     }
 
-    if(tokenVerify(token_public)){
-
-        const db = new Client(pgInit)
-        db.connect((err)=>{
-            if(err) {
-                console.log(err)
-            }
-        })
-
-        const sql="DELETE FROM badonnaproject.reply WHERE reply_num=$1;"
-        const values=[reply_number]
-        
-        db.query(sql,values,(err,row)=>{
-
-            if(!err){
-                result.success=true
-            }else{
-                console.log(err)
-            }
-            //로깅 남기기
-            logFuntion(api_name,req_host, req_data, row.rows[0],api_call_time)
-
+    try{
+        if(reply_number.length == 0 || reply_number == null || Number.isInteger(reply_number)){
+            result.message="옳바르지 않은 대댓글 번호 입니다."
             res.send(result)
-            db.end()
-            
-        })
+        }else{
 
+            if(tokenVerify(token_public)){
+
+                const db = new Client(pgInit)
+                db.connect((err)=>{
+                    if(err) {
+                        console.log(err)
+                    }
+                })
+
+                const sql="DELETE FROM badonnaproject.reply WHERE reply_num=$1;"
+                const values=[reply_number]
+                
+                db.query(sql,values,(err,row)=>{
+
+                    if(!err){
+                        result.success=true
+                    }else{
+                        console.log(err)
+                    }
+                    //로깅 남기기
+                    logFuntion(api_name,req_host, req_data, row.rows[0],api_call_time)
+
+                    res.send(result)
+                    db.end()
+                    
+                })
+
+            }else{
+                result.message="잘못된 토큰"
+                res.send(result)
+            }
+        }
+
+    }catch(e){
+        result.message="에러 입니다."
+        res.send(result)
     }
 
 })
